Add WindPointList to IsobaricRepo, making it possible to expose only one list of WindPoints instead of always needing every list to get all the information of a PressureLayer

+++ app/src/main/java/no/uio/ifi/in2000/team_17/data/isobaricgrib/model/WindPoint.kt
@@ -6,7 +6,7 @@ data class WindPoint(
     val windSpeed: Double = 0.0,
     val windFromDirection: Double = 0.0,
     val temperature: Double = 0.0,
-    val pressure: Int = 1000,
+    val pressure: Double = 1000.00,
     val height: Double = 0.0
 )
